scrapper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
redis==5.0.1
python-dotenv==1.0.0
urllib3==2.0.7
lxml==4.9.3
